changes in collitions and pointers

+++ codigo/src/concepts/collition/resolve_collition/sq_sq.cpp
@@ -4,7 +4,7 @@
 
 void resolve_collition (Square& sq1, Square& sq2) {
   AngDir2 diff = sq1.position - sq2.position;
-  Dir2 size = Dir2 {(float)sq1.width + sq2.width, (float)sq2.height + sq1.height};
+  Dir2 size = Dir2 {sq1.width + sq2.width, sq2.height + sq1.height};
 
   int pos = gti (absv(diff.y) + size.x, absv(diff.x) + size.y);
   float * vel_elem1 = &sq1._velocity.x + pos;
